PATCH: binary_search use wrapping to allow overflow
