+++ 1_slices_modules_and_objects/exercises/binary_search/src/lib.rs
@@ -1,21 +1,23 @@
+use std::num::Wrapping;
+
 /// Finds the index of `value` in the sorted slice `v`
 /// Returns the index if found, None otherwise
 pub fn binary_search(v: &[i32], value: i32) -> Option<usize> {
-    let mut high: i32 = v.len() as i32 - 1;
+    let mut high = Wrapping(v.len() as i32 - 1);
     let mut low = 0;
     
-    while high >= low {
-        let mid = ((high - low)) / 2 + low;
-        let mid_index = mid as usize;
+    while high.0 >= low {
+        let mid = Wrapping(((high.0 - low)) / 2 + low);
+        let mid_index = mid.0 as usize;
         let cur = v[mid_index];
 
         if cur == value {
             return Some(mid_index);
         }
         if cur < value {
-            low = mid + 1;
+            low = mid.0 + 1;
         } else {
-            high = mid - 1;
+            high = mid - Wrapping(1);
         }
     }
